Refactor GuessItem: Move feedback logic to separate file and improve feedback handling Implement useLandingPageLogic hook: Add smooth scrolling and mission text management Remove unused logger from PolidleApiService
